requirements.txt
flask 
flask-pymongo 
authlib 
python-dotenv 
requests
